Format file

+++ US_Election.Dal/Models/VoteViewModel.cs
@@ -6,7 +6,7 @@
 
         public string CandidateName { get; set; }
 
-        public double Parcentage { get; set; }
+        public float Parcentage { get; set; }
 
         public int CandidateId { get; set; }
